Rename package server to simulator
