+++ simulator/add_object.go
@@ -1,4 +1,4 @@
-package server
+package simulator
 
 import (
 	"strings"
