+++ simulator/inform_test.go
@@ -1,4 +1,4 @@
-package server
+package simulator
 
 import (
 	"testing"
