+++ simulator/metrics/noop.go
